add data processing + style fix typescript
